+++ src/status/condition/mod.rs
@@ -1,0 +1,2 @@
+pub mod daemonset;
+pub mod statefulset;
